+++ src/common/utils/getAppCtx.ts
@@ -1,5 +1,5 @@
 import { UserService } from "@/api/user/user.service";
-import type { AppCtx } from "../models/AppContext";
+import type { AppCtx } from "../models/appCtx";
 
 const productionAppCtx: AppCtx = {
 	userService: new UserService(),
